requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2